requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
anyio==3.6.2
Appium-Python-Client==2.10.1
async-generator==1.10
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
genson==1.2.2
greenlet==2.0.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
loguru==0.6.0
MarkupSafe==2.1.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
parameterized==0.8.1
playwright==1.33.0
pluggy==1.0.0
py==1.11.0
pyee==9.0.4
PyMySQL==1.0.3
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-playwright==0.3.3
pytest-rerunfailures==11.1.2
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
pytest-twisted==1.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
seldom==3.2.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.6.1
urllib3==2.0.2
webdriver-manager==3.8.5
wsproto==1.2.0
xtestrunner==1.7.0
zope.interface==6.0
